pyproject.toml
[project]
name = "spt-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "dotenv>=0.9.9",
    "google-genai>=1.49.0",
    "gtts>=2.5.4",
    "sounddevice>=0.5.3",
    "whisper>=1.1.10",
]
